Refactor js to use modules and follow best practices
